requirements.txt
requests
scipy
statsmodels
scikit-learn
pandas
numpy
seaborn
matplotlib
python-dotenv
ipykernel
beautifulsoup4
selenium
discord.py
openpyxl
altair